// src/app/interfaces/cinema.ts
export interface CinemaCreation {
    name: string;
    latitude: number;
    longitude: number;
}

export interface CinemaView {
    name: string;
    latitude: number;
    longitude: number;
}

export interface CinemaList {
    id: number;
    name: string;
    latitude: number;
    longitude: number;
}